requirements.txt
# Content Recycler — Python Dependencies
# Install: pip install -r requirements.txt

# Core RAG / LLM
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
google-generativeai>=0.7.0

# Vector DB
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.7.0

# Data
pandas>=2.0.0
numpy>=1.26.0

# Evaluation
sacrebleu>=2.4.0
nltk>=3.8.0

# UI
streamlit>=1.35.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
